+++ src/business/UserBusiness.ts
@@ -1,7 +1,7 @@
 import { USER_ROLES, User, toUserRole } from "../model/User";
 import { IdGenerator } from "../services/IdGenerator";
 import HashManager from "../services/HashManager";
-import { UserDatabase } from "../data/UserDatabase";
+import { UserDatabase } from "../data/UserDataBase";
 import { Authenticator } from "../services/Authenticator";
 
 export class UserBusiness {
