+++ src/controllers/FriendController.ts
@@ -1,8 +1,6 @@
 import { BaseDatabase } from "../data/BaseDatabase";
-import { UsersRelationDatabase } from "../data/UsersRelationDatabase";
-import { UserDatabase } from "../data/UserDatabase";
-import { Authenticator } from "../services/Authenticator";
 import { Request, Response } from "express";
+import { FriendBusiness } from "../business/FriendBusiness";
 
 
 export class FriendController {
@@ -11,26 +9,7 @@ export class FriendController {
             const token = req.headers.authorization as string;
             const addFriendById = req.body.addFriendById;
 
-            const authenticator = new Authenticator();
-            const authenticationData = authenticator.getData(token);
-            const userId = authenticationData.id;
-
-            if (!addFriendById) {
-                throw new Error("Insira um id de usuário válido")
-            };
-
-            const userDatabase = new UserDatabase();
-            const user = await userDatabase.getUserById(addFriendById);
-
-            if (!user) {
-                throw new Error("Usuário não existe")
-            };
-
-            const usersRelationDatabase = new UsersRelationDatabase();
-            await usersRelationDatabase.addFriend(
-                userId,
-                addFriendById
-            );
+            await new FriendBusiness().addFriend(token, addFriendById);
 
             res.status(200).send({ message: "Usuário adicionado com sucesso! :)" });
 
@@ -45,26 +24,7 @@ export class FriendController {
             const token = req.headers.authorization as string;
             const undoFriendshipById = req.body.undoFriendshipById;
 
-            const authenticator = new Authenticator();
-            const authenticationData = authenticator.getData(token);
-            const userId = authenticationData.id;
-
-            if (!undoFriendshipById) {
-                throw new Error("Insira um id de usuário válido")
-            };
-
-            const userDatabase = new UserDatabase();
-            const user = await userDatabase.getUserById(undoFriendshipById);
-
-            if (!user) {
-                throw new Error("Usuário não existe")
-            };
-
-            const usersRelationDatabase = new UsersRelationDatabase();
-            await usersRelationDatabase.undoFriendship(
-                userId,
-                undoFriendshipById
-            );
+            await new FriendBusiness().undoFriendship(token, undoFriendshipById);
 
             res.status(200).send({ message: "Você deixou de seguir o usuário com sucesso!" });
